lots of new stuff for ajax.php. Working "loading" function on sellLoot.php that loads journal data via simple Ajax. I quite like working with jQuery; never really done it before.
Added a new degub options since I was getting wierd results from the API and I need to investigate further. It seems that when you first call the API after
new data is avaialble (after you sell stuff) the order in which it shows up is not constant. It's only when you load it again for a second time or more that
it seems transactions are arranged in a sane order. I will most likely have to do my own array sorting based off transaction id (or time), but to be sure
that's the problem I'm adding these debug values
problem